tech: Orders endpoint

+++ db/migrate/20210115181915_create_order_payment.rb
@@ -2,7 +2,7 @@ class CreateOrderPayment < ActiveRecord::Migration[6.1]
   def change
     create_table :order_payments do |t|
       t.references :order, null: false, foreign_key: true
-      t.integer :external_code
+      t.bigint :external_code
       t.integer :payer_id
       t.integer :installments
       t.string :payment_type
