+++ db/migrate/20210115182323_create_buyer.rb
@@ -1,7 +1,7 @@
 class CreateBuyer < ActiveRecord::Migration[6.1]
   def change
     create_table :buyers do |t|
-      t.integer :external_code
+      t.bigint :external_code
       t.string :nickname
       t.string :email
       t.string :first_name
