+++ db/migrate/20210115184238_create_order_shipping.rb
@@ -2,7 +2,7 @@ class CreateOrderShipping < ActiveRecord::Migration[6.1]
   def change
     create_table :order_shippings do |t|
       t.references :order, null: false, foreign_key: true
-      t.integer :external_code
+      t.bigint :external_code
       t.string :shipment_type
       t.datetime :date_created
